# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HRUO Enterprise HCM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance/records:
    get:
      operationId: listAttendanceRecords
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: clockIn
      tags: [attendance]
      summary: Clock in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/records/{id}/clock-out:
    post:
      operationId: clockOut
      tags: [attendance]
      summary: Clock out
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/summary:
    get:
      operationId: getAttendanceSummary
      tags: [attendance]
      summary: Get attendance summary
      parameters:
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"

  # ── PAYROLL ───────────────────────────────────────────────────────────────
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List shared documents
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload / create document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  # ── INCIDENTS ─────────────────────────────────────────────────────────────
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incidents
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Report incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"

  # ── ANNOUNCEMENTS ─────────────────────────────────────────────────────────
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

  # ── DISCUSSIONS ───────────────────────────────────────────────────────────
  /discussions:
    get:
      operationId: listDiscussions
      tags: [discussions]
      summary: List discussions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discussion"
    post:
      operationId: createDiscussion
      tags: [discussions]
      summary: Create discussion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscussionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discussion"

  # ── TASKS ─────────────────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: assigneeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── ANALYTICS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AttendanceRecord:
      type: object
      required: [id, employeeId, date]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        date:
          type: string
        clockIn:
          type: ["string", "null"]
        clockOut:
          type: ["string", "null"]
        totalHours:
          type: ["number", "null"]
        overtime:
          type: ["number", "null"]
        status:
          type: string
          enum: [present, absent, late, half_day, holiday]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        ipAddress:
          type: ["string", "null"]
    ClockInput:
      type: object
      required: [employeeId]
      properties:
        employeeId:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        ipAddress:
          type: string
        selfieUrl:
          type: string
    AttendanceSummary:
      type: object
      required: [month, totalPresent, totalAbsent, totalLate, totalOvertime]
      properties:
        month:
          type: string
        totalPresent:
          type: integer
        totalAbsent:
          type: integer
        totalLate:
          type: integer
        totalOvertime:
          type: number
        byEmployee:
          type: array
          items:
            type: object
            properties:
              employeeId:
                type: integer
              employeeName:
                type: string
              presentDays:
                type: integer
              absentDays:
                type: integer

    # ── PAYROLL ───────────────────────────────────────────────────────────────
...
    Document:
      type: object
      required: [id, title, category]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        sharedWith:
          type: string
        uploadedBy:
          type: ["string", "null"]
        createdAt:
          type: string
    DocumentInput:
      type: object
      required: [title, category]
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        fileUrl:
          type: string
        sharedWith:
          type: string

    # ── INCIDENTS ─────────────────────────────────────────────────────────────
    Incident:
      type: object
      required: [id, employeeId, type, date, status]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        type:
          type: string
        date:
          type: string
        description:
          type: ["string", "null"]
        actionTaken:
          type: ["string", "null"]
        penalty:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, approved, closed]
        createdAt:
          type: string
    IncidentInput:
      type: object
      required: [employeeId, type, date]
      properties:
        employeeId:
          type: integer
        type:
          type: string
        date:
          type: string
        description:
          type: string
        actionTaken:
          type: string
        penalty:
          type: number
    IncidentUpdate:
      type: object
      properties:
        status:
          type: string
        actionTaken:
          type: string
        penalty:
          type: number

    # ── ANNOUNCEMENTS ─────────────────────────────────────────────────────────
    Announcement:
      type: object
      required: [id, title, published]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: ["string", "null"]
        published:
          type: boolean
        announcementDate:
          type: ["string", "null"]
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        content:
          type: string
        published:
          type: boolean
        announcementDate:
          type: string
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        published:
          type: boolean

    # ── DISCUSSIONS ───────────────────────────────────────────────────────────
    Discussion:
      type: object
      required: [id, title, authorId]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: ["string", "null"]
        authorId:
          type: integer
        authorName:
          type: ["string", "null"]
        scope:
          type: string
        replyCount:
          type: integer
        createdAt:
          type: string
    DiscussionInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        content:
          type: string
        scope:
          type: string

    # ── TASKS ─────────────────────────────────────────────────────────────────
    Task:
      type: object
      required: [id, title, status, priority]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [todo, in_progress, review, done]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: string
        priority:
          type: string
        status:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: string
        priority:
          type: string
        status:
          type: string

    # ── ANALYTICS ─────────────────────────────────────────────────────────────
...